requirements.txt
os
time
random
re
markdown
keyboard
SpeechRecognition
google-generativeai
gTTS
playsound
